test: refactor file structure
