+++ tests/recovery/recovery_test/main_test.go
@@ -13,6 +13,8 @@ import (
 
 const (
 	dataSubPath = "recovery-data"
+	dirPath     = "kopia_dummy_repo"
+	dataPath    = "crash-consistency-data"
 )
 
 var repoPathPrefix = flag.String("repo-path-prefix", "", "Point the robustness tests at this path prefix")
